Improved handling of Superfeedr HTML Fragment Subscriptions
Previously SuperfeedrHubs would just subscribe to the whole HTML of a page, but in practise
that results in huge quanitites of unnecessary pings. This update limits the default subscription
to a CSS selector (by default '.h-entry'), and makes sure that the whole page content is
fetched on pings to ensure correct whole-content parsing (the pings only contain the fragment).